Add tvOS to components (#13)
This makes the components library available on tvOS. It seemed the the
default package.swift was implicitly using a minimum tvOS version of 14,
which conflicted with our core SDK at 17. A swift 5.9 file is needed to
set it to 17 explicitly here.

I added visionOS explicitly as well just for clarity, but it wasn't
strictly necessary to do.

I was able to test this in the simulator and used our components
quickstart guide to join a meet call and show video.